+++ api/src/resources/TimeClock.ts
@@ -1,6 +1,7 @@
-import { Request, Response, NextFunction } from 'express';
+import { Request, Response, NextFunction } from 'express'
 import { SingleTable } from '../SingleTable'
-import { Database } from '../Database';
+import { Database } from '../Database'
+import { IRequest } from '../UserSession'
 import * as Logger from 'bunyan'
 
 export class TimeClock extends SingleTable {
@@ -33,8 +34,8 @@ export class TimeClock extends SingleTable {
 
   public async read (req: Request, res: Response, next: NextFunction) {
     this.logger.trace (`timeclock.read(${[req.params.user_id,req.params.date]})`)
-    const sql = `SELECT * FROM time_clock WHERE user_id = ? AND date = ?`
     try {
+      const sql = `SELECT * FROM time_clock WHERE user_id = ? AND date = ?`
       let row = await this.db.get (sql, [req.params.user_id, req.params.date])
       row ? res.status (200).json (row) : res.status (404).json ()
       next ()
@@ -46,8 +47,9 @@ export class TimeClock extends SingleTable {
 
   public async update (req: Request, res: Response, next: NextFunction) {
     this.logger.trace (`timeclock.update(${[req.params.user_id,req.params.date]})`)
-    let status = 400
     try {
+      this.validateAccess (<IRequest>req)
+      let status = 400
       let [ names, values ] = this.sqlGenerator.buildParameterList (req.body)
       if (names.length > 0) {
         const sql = `UPDATE time_clock AS tc SET ` + this.getUpdateList(names) + ' WHERE tc.user_id = ? AND tc.date = ?'
@@ -67,6 +69,7 @@ export class TimeClock extends SingleTable {
   public async delete (req: Request, res: Response, next: NextFunction) {
     this.logger.trace (`timeclock.update(${[req.params.user_id,req.params.date]})`)
     try {
+      this.validateAccess (<IRequest>req)
       const result: any = await this.db.run (`DELETE FROM time_clock WHERE user_id = ? AND date = ?`, [req.params.user_id, req.params.date])
       result && result.changes > 0 ? res.status (204).json () : res.status (404).json ()
       next ()
@@ -76,13 +79,26 @@ export class TimeClock extends SingleTable {
     }
   }
 
-  public validateCreate(req: Request) {
+  public validateCreate (req: Request) {
     if (!req.body.date) {
       throw new Error ('400:Create new time clock entry: date is required')
     }
     if (!req.body.user_id) {
       throw new Error ('400:Create new time clock entry: user_id is required')
     }
+  }
+
+  public validateAccess (req: IRequest) {
+    if (!this.isWriteAccess (req) || req.session.hasAdmin ()) {
+      return
+    }
+    if (req.session.hasUserId (Number (req.params.user_id))) {
+      return
+    }
+    if (req.method === 'POST' && req.body && req.body && req.session.hasUserId (Number (req.body.user_id))) {
+      return
+    }
+    throw new Error ('403:Insufficient credentials')
   }
 }
